tests/it/grpc_client/requirements.txt
# Direct dependencies:
#  grpcio protobuf
grpcio==1.54.2
protobuf==4.23.3
